+++ auto-kol/agent/.env.sample
@@ -29,3 +29,6 @@ CHROMA_URL=http://localhost:8000
 
 # AutoDrive Configuration
 DSN_API_KEY=your_dsn_api_key
+
+# CORS Configuration
+CORS_ORIGINS=http://localhost:3000
