+++ auto-kol/agent/src/api/middleware/cors.ts
@@ -1,0 +1,22 @@
+import cors from 'cors';
+import { config } from '../../config/index.js';
+
+const allowedOrigins = config.CORS_ORIGINS?.split(',') || ['http://localhost:3000'];
+
+export const corsMiddleware = cors({
+    origin: (origin, callback) => {
+        if (!origin) {
+            return callback(null, true);
+        }
+
+        if (allowedOrigins.indexOf(origin) !== -1 || config.NODE_ENV === 'development') {
+            callback(null, true);
+        } else {
+            callback(new Error('Not allowed by CORS'));
+        }
+    },
+    credentials: true,
+    methods: ['GET', 'POST', 'PUT', 'DELETE', 'OPTIONS'],
+    allowedHeaders: ['Content-Type', 'Authorization'],
+    maxAge: 86400 // 24 hours
+});
