+++ auto-kol/agent/src/api/routes/health.ts
@@ -1,0 +1,9 @@
+import { Router } from 'express';
+
+const router = Router();
+
+router.get('/health', (_, res) => {
+    res.json({ status: 'ok' });
+});
+
+export default router;
